aethercast/aims_service/requirements.txt
Flask
python-dotenv
openai
